Fixed some spacing related issues
