+++ pnparser/templates/designer.py
@@ -3,7 +3,7 @@ import json as Json
 import re as Regex
 
 ICONS_DB: str = Path.join(Path.dirname(__file__), "../data/designer_icons.json")
-TEMPLATE: str = "[[File:{0}|20px|link=]] {1}<br>"
+TEMPLATE: str = "[[File:{0}|20px|link=]] {1}<br>\n"
 
 
 class Designer:
